Consistently use Z_NULL to initialize the opaque object.
There's also an actual benefit to this. When using the unmodified zconf.h.in,
compress.c and uncompr.c end up without a definition for NULL in both GCC and
Clang, when compiling for the x86_64-unknown-linux-gnu target. Using Z_NULL
here allows zlib-ng (without the gzip parts) to be compiled under this setup
by just symlinking zconf.h.in to zconf.h, without having to run ./configure or
CMake before, which is especially useful for projects that statically link
zlib-ng or use Git submodules.