Add detection for 64bit long int to support padding
Zlib-ng introduces a 32bit padding member in its data
struct to make zlib-ng compatible with zlib, without
which there will be runtime errors when making zlib-ng
a binary replacement of zlib. Now the padding is enabled
only when both ZLIB_COMPAT and X86_64 are defined, so it
needs support for other platforms where sizeof(long int)
is not equal to sizeof(int).

This patch fixes the problem by adding the detection for
64bit long int.

Signed-off-by: Richael Zhuang <richael.zhuang@arm.com>