Optimize inflate_fast for a 0.8% speedup
When inflate_fast checks for extra length bits, it first checks if the
number of extra length bits (in op) is non-zero. However, if the number
of extra length bits is 0, the `bits < op` check will be false, BITS(op)
will be 0, and DROPBITS(op) will do nothing. So, drop the conditional,
for a speedup of about 0.8%.

This makes the handling of extra length bits match the handling of extra
dist bits, which already lacks the extra conditional.