.gitattributes: Do no longer mess with line-endings and leave as is
Although Git is able to automatically modify line-endings during checkin
and checkout, this brings a lot of trouble, especially when trying to
use a repository from different platforms (as Windows and Linux).
Therefore, Git should leave line-endings as is and should not
automatically modify them during checkin and checkout.

It is the committer's responsibility to commit files with correct
line-endings (preferably LF) and the reviewer's responsibility to ensure
that no wrong line-endings are introduced through pull-requests.

Probably, some tool like clang-format or an .editorconfig file for
text-editors and IDEs should enforce line-endings.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>