Disable MSVC warning 4819
Currently, zlib-ng doesn't compile on my machine due to warning C4819: The file contains a character that cannot be represented in the current code page (936). Save the file in Unicode format to prevent data loss.

The non-ascii characters live in zbuild.h.
```
 * gcc says: "warning: implicit declaration of function ‘...’"
 * g++ says: "error: new declaration ‘...’ ambiguates built-in declaration ‘...’"
```

Because zlib-ng treats warning as error, the project doesn't compile. If
this warning is disabled, zlib-ng compiles well.